integration_service/requirements.txt
kafka-python
zeep
requests
